pyproject.toml
[tool.poetry]
name = "komtek"
version = "0.1.0"
description = ""
authors = ["Gregory <natsage.g@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
model-mommy = "^2.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
